email_sending_service/requirements.txt
alembic==1.9.3
Faker==16.8.1
flake8==6.0.0
greenlet==2.0.2
importlib-metadata==6.0.0
importlib-resources==5.10.2
Mako==1.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
pycodestyle==2.10.0
pyflakes==3.0.1
python-dateutil==2.8.2
python-dotenv==0.21.1
six==1.16.0
SQLAlchemy==2.0.3
typing_extensions==4.4.0
zipp==3.13.0
